+++ server/index.js
@@ -1,9 +1,7 @@
 import mongoose from 'mongoose';
 import colors from 'colors';
 import dotenv from 'dotenv';
-
 import app from './app.js';
-
 dotenv.config();
 
 const PORT = process.env.PORT || 5000;
@@ -14,11 +12,13 @@ app.listen(PORT, () => {
   );
 });
 
+//DB connection
 mongoose
   .connect(process.env.MONGODB_URL, {
     useCreateIndex: true,
     useNewUrlParser: true,
     useUnifiedTopology: true,
+    useFindAndModify: false,
   })
   .then(() => console.log('MongoDB connected succesfully'.cyan.underline))
   .catch((err) => console.log('Error'.red.underline.bold));
